[module_icon_setup] remove modules setup page
