+++ frappe/config/setup.py
@@ -37,13 +37,6 @@ def get_data():
 					"description": _("Set Permissions on Document Types and Roles")
 				},
 				{
-					"type": "page",
-					"name": "modules_setup",
-					"label": _("Show / Hide Modules"),
-					"icon": "fa fa-upload",
-					"description": _("Show or hide modules globally.")
-				},
-				{
 					"type": "doctype",
 					"name": "Role Permission for Page and Report",
 					"description": _("Set custom roles for page and report")
