--- frappe/core/page/modules_setup/README.md
@@ -1,1 +1,0 @@
-Interface to show / hide modules.
