--- frappe/core/page/modules_setup/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
-# MIT License. See license.txt
-
