--- frappe/core/page/modules_setup/modules_setup.html
@@ -1,47 +1,0 @@
-<!-- jinja -->
-{% if users %}
-<div class="padding modules-setup">
-	<div class="row">
-
-		<!-- TAKE -->
-		<div class="col-sm-3" name="set_for">
-			<select class="form-control" name="setup_for">
-				<option value="everyone">{{ _("For Everyone") }}</option>
-				<option value="user" selected>{{ _("For User") }}</option>
-			</select>
-		</div>
-
-
-		<!-- TAKE -->
-		<div class="col-sm-3">
-			<select class="form-control" name="user">
-				{% for user in users %}
-				<option value="{{ user.name | e }}"
-					{% if user.name == frappe.user %}selected{% endif %}>
-					<!-- {{ variable | e }} "e" or "escape(s)" will escape the characters such "<, >, &, '"
-					in the HTML text (http://jinja.pocoo.org/docs/dev/templates/#escape) -->
-					{{ (user.first_name or "") | e }} {{ (user.last_name or "") | e }} ({{ user.name | e }})</option>
-				{% endfor %}
-			</select>
-		</div>
-
-
-		<div class="col-sm-3">
-			  <div class="checkbox" style="margin-top: 5px;">
-			    <label>
-			      <input type="checkbox" class="check-all">
-					{{ _("Select All") }}
-			    </label>
-			  </div>
-		</div>
-	</div>
-
-	<!-- TAKE -->
-	<div class="block-warning hidden">
-		<div class="alert alert-warning small" style="margin: 0px; margin-top: 15px;">{{ _("Global Settings: Users will only be able to choose checked icons") }}</div>
-	</div>
-
-
-</div>
-{% endif %}
-<div class="padding modules-setup-icons"></div>
