--- frappe/core/page/modules_setup/modules_setup.json
@@ -1,20 +1,0 @@
-{
- "content": null, 
- "creation": "2012-10-04 18:45:29", 
- "docstatus": 0, 
- "doctype": "Page", 
- "icon": "fa fa-cog", 
- "idx": 1, 
- "modified": "2017-04-12 18:44:53.283640", 
- "modified_by": "Administrator", 
- "module": "Core", 
- "name": "modules_setup", 
- "owner": "Administrator", 
- "page_name": "modules_setup", 
- "roles": [], 
- "script": null, 
- "standard": "Yes", 
- "style": null, 
- "system_page": 1, 
- "title": "Modules Setup"
-}
